chore: cleans up action structure
