+++ src/helpers/actions.ts
@@ -1,4 +1,5 @@
 import { Page } from 'playwright-core';
+import { Step, WalletOptions } from '../wallets/wallets';
 import { getElementByContent, getInputByLabel } from './selectors';
 
 export const waitForChromeState = async (page: Page): Promise<void> => {
@@ -47,3 +48,35 @@ export const typeOnInputField = async (
   await input.type(text);
   return true;
 };
+
+export const performPopupAction = async (page: Page, action: (popup: Page) => Promise<void>): Promise<void> => {
+  const popup = await page.context().waitForEvent('page'); // Wait for the popup to show up
+
+  await action(popup);
+  if (!popup.isClosed()) await popup.waitForEvent('close');
+};
+
+export const performSidepanelAction = async (page: Page, action: (sidepanel: Page) => Promise<void>): Promise<void> => {
+  const sidepanel = page
+    .context()
+    .pages()
+    .find((p) => p.url().includes('sidepanel.html'));
+
+  if (!sidepanel) {
+    const pageUrls = page
+      .context()
+      .pages()
+      .flatMap((p) => p.url());
+    throw new Error('Sidepanel page not found. Current pages:\n' + pageUrls.join('\n- '));
+  }
+
+  await action(sidepanel);
+};
+
+export const performSetup =
+  (page: Page, defaultSteps: Step<WalletOptions>[]) =>
+  async <Options = WalletOptions>(options?: Options, steps: Step<Options>[] = defaultSteps): Promise<void> => {
+    for (const step of steps) {
+      await step(page, options);
+    }
+  };
