--- src/wallets/metamask/actions/addNetwork.ts
@@ -1,47 +1,0 @@
-import { Page } from 'playwright-core';
-import { clickOnButton, clickOnElement, waitForChromeState } from '../../../helpers';
-import { AddNetwork } from '../../../types';
-
-import { getErrorMessage, openNetworkDropdown } from './helpers';
-import { switchNetwork } from './switchNetwork';
-
-export const addNetwork =
-  (page: Page) =>
-  async ({ networkName, rpc, chainId, symbol }: AddNetwork): Promise<void> => {
-    await openNetworkDropdown(page);
-    await clickOnElement(page, 'Custom');
-    await clickOnButton(page, 'Add custom network');
-
-    await page.getByTestId('network-form-network-name').fill(networkName);
-    await page.getByTestId('test-add-rpc-drop-down').click();
-    await clickOnButton(page, 'Add RPC URL');
-    await page.getByTestId('rpc-url-input-test').fill(rpc);
-    await clickOnButton(page, 'Add URL');
-    await page.getByTestId('network-form-chain-id').fill(String(chainId));
-    await page.getByTestId('network-form-ticker-input').fill(symbol);
-
-    const errorMessage = await getErrorMessage(page);
-    if (errorMessage) {
-      await page.getByRole('dialog').getByRole('button', { name: 'Close' }).click();
-      throw new SyntaxError(errorMessage);
-    }
-
-    await clickOnButton(page, 'Save');
-    await page.getByTestId('modal-header-close-button').click();
-
-    // This popup is fairly random in terms of timing
-    // and can show before switch to network click is gone
-    const gotItClick = (): Promise<void> =>
-      page.waitForTimeout(2000).then(() =>
-        page
-          .locator('button', { hasText: 'Got it' })
-          .isVisible()
-          .then((gotItButtonVisible) => {
-            if (gotItButtonVisible) return clickOnButton(page, 'Got it');
-            return Promise.resolve();
-          }),
-      );
-
-    await waitForChromeState(page);
-    await Promise.all([switchNetwork(page)(networkName), gotItClick()]);
-  };
