+++ src/wallets/metamask/actions/helpers/selectors.ts
@@ -1,4 +1,4 @@
-import { ElementHandle, Page } from 'playwright-core';
+import { ElementHandle, Locator, Page } from 'playwright-core';
 
 export const getSettingsSwitch = (page: Page, text: string): Promise<ElementHandle | null> =>
   page.waitForSelector([`//span[contains(.,'${text}')]/parent::div/following-sibling::div/label/div`].join('|'));
@@ -10,4 +10,21 @@ export const getErrorMessage = async (page: Page): Promise<string | undefined> =
   } catch (_) {
     return undefined;
   }
+};
+
+export const accountList = (page: Page): Locator => {
+  return page.locator('.multichain-account-cell');
+};
+
+export const accountListItem = (page: Page, name: string): Locator => {
+  return accountList(page).filter({ has: page.getByText(name, { exact: true }) });
+};
+
+const networkMenuItemRegex = /network-list-item-eip\d+:\d+$/;
+const networkList = (page: Page): Locator => {
+  return page.getByTestId(networkMenuItemRegex);
+};
+
+export const networkListItem = (page: Page, name: string): Locator => {
+  return networkList(page).filter({ has: page.getByText(name, { exact: true }) });
 };
