--- src/wallets/metamask/actions/sign.ts
@@ -1,11 +1,0 @@
-import { Page } from 'playwright-core';
-import { performSidepanelAction } from './util';
-
-export const sign = (page: Page) => async (): Promise<void> => {
-  await performSidepanelAction(page, async (popup) => {
-    await popup.bringToFront();
-    await popup.reload();
-
-    await popup.getByTestId('confirm-footer-button').click();
-  });
-};
