--- src/wallets/metamask/actions/unlock.ts
@@ -1,13 +1,0 @@
-import { Page } from 'playwright-core';
-import { closePopup } from '../setup/setupActions';
-
-export const unlock =
-  (page: Page) =>
-  async (password = 'password1234'): Promise<void> => {
-    await page.bringToFront();
-
-    await page.getByTestId('unlock-password').fill(password);
-    await page.getByTestId('unlock-submit').click();
-
-    await closePopup(page);
-  };
